迁移主题设置 SettingState
