--- app/src/main/kotlin/indi/dmzz_yyhyy/lightnovelreader/ui/home/settings/theme/ThemeSettingStates.kt
@@ -1,32 +1,0 @@
-package indi.dmzz_yyhyy.lightnovelreader.ui.home.settings.theme
-
-import android.net.Uri
-import androidx.compose.runtime.getValue
-import androidx.compose.ui.graphics.Color
-import indi.dmzz_yyhyy.lightnovelreader.data.setting.AbstractSettingState
-import indi.dmzz_yyhyy.lightnovelreader.data.userdata.UserDataPath
-import indi.dmzz_yyhyy.lightnovelreader.data.userdata.UserDataRepository
-import kotlinx.coroutines.CoroutineScope
-
-class ThemeSettingState(
-    userDataRepository: UserDataRepository,
-    coroutineScope: CoroutineScope
-) : AbstractSettingState(coroutineScope) {
-    val darkModeKeyUserData = userDataRepository.stringUserData(UserDataPath.Settings.Display.DarkMode.path)
-    val dynamicColorsKeyUserData = userDataRepository.booleanUserData(UserDataPath.Settings.Display.DynamicColors.path)
-    val appLocaleKeyUserData = userDataRepository.stringUserData(UserDataPath.Settings.Display.AppLocale.path)
-    val enableBackgroundImageUserData = userDataRepository.booleanUserData(UserDataPath.Reader.EnableBackgroundImage.path)
-    val backgroundImageDisplayModeUserData = userDataRepository.stringUserData(UserDataPath.Reader.BackgroundImageDisplayMode.path)
-    val backgroundImageUriUserData = userDataRepository.uriUserData(UserDataPath.Reader.BackgroundImageUri.path)
-    val backgroundDarkImageUriUserData = userDataRepository.uriUserData(UserDataPath.Reader.BackgroundDarkImageUri.path)
-    val textColorUserData = userDataRepository.colorUserData(UserDataPath.Reader.TextColor.path)
-
-    val darkModeKey by darkModeKeyUserData.asState("FollowSystem")
-    val dynamicColorsKey by dynamicColorsKeyUserData.asState(false)
-    val appLocaleKey by appLocaleKeyUserData.asState("zh-CN")
-    val enableBackgroundImage by enableBackgroundImageUserData.safeAsState(false)
-    val backgroundImageDisplayMode by backgroundImageDisplayModeUserData.safeAsState("fixed")
-    val backgroundImageUri by backgroundImageUriUserData.safeAsState(Uri.EMPTY)
-    val backgroundDarkImageUri by backgroundDarkImageUriUserData.safeAsState(Uri.EMPTY)
-    val textColor by textColorUserData.safeAsState(Color.Unspecified)
-}
